Add test abstraction (can be used for simple taks, like task and one answer) Start to make TestManager
