+++ DataModel/DataModel.csproj
@@ -58,7 +58,10 @@
     <Compile Include="Programs\AbstractProgram.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Support\Singletone.cs" />
-    <Compile Include="Tests\AbstractTest.cs" />
+    <Compile Include="TestsManager.cs" />
+    <Compile Include="Tests\Test.cs" />
+    <Compile Include="Tests\ITestable.cs" />
+    <Compile Include="Tests\TestItems\SimpleTestItem.cs" />
     <Compile Include="Users\AbstractUser.cs" />
     <Compile Include="Users\Root.cs" />
   </ItemGroup>
