+++ DataModel/Tests/ITestable.cs
@@ -1,0 +1,13 @@
+﻿namespace DataModel.Tests
+{
+    public interface ITestableItem
+    {
+        
+        double Grade { get; }
+        
+        double MaxGrade { get; }
+        
+        
+        
+    }
+}
