+++ DataModel/TestsManager.cs
@@ -1,0 +1,12 @@
+﻿using DataModel.Support;
+
+namespace DataModel
+{
+    public class TestsManager:Singletone<TestsManager>
+    {
+        
+        
+        
+        
+    }
+}
